Added page (Model Accuracy Analysis) to update run results table to indicate if the onset was used to create the model.
This page will be developed to allow checking if the model is performing
well on instances that it was not trained with.  Models will normlly be
created in Weka using 10 fold cross validation, so overfitting should
not be happening - but this page will help confirm that - by using the
other page 'Manually evaluate Weka Model result' to create some
new/unseen by the model creation process, that can then be analysed.

Also the java code - Main4 was updated to use the new weka model type.